Update Error Avatar

+++ app/src/main/java/com/imtuc/week4retrofit/model/MovieDetails.kt
@@ -1,5 +1,7 @@
 package com.imtuc.week4retrofit.model
 
+import java.math.BigInteger
+
 data class MovieDetails(
     val adult: Boolean,
     val backdrop_path: String,
@@ -17,7 +19,7 @@ data class MovieDetails(
     val production_companies: List<ProductionCompany>,
     val production_countries: List<ProductionCountry>,
     val release_date: String,
-    val revenue: Int,
+    val revenue: BigInteger,
     val runtime: Int,
     val spoken_languages: List<SpokenLanguage>,
     val status: String,
